requirements.txt
Werkzeug==2.1.2
flask==2.1.2
line-bot-sdk==2.3.0
